requirements.txt
asgiref==3.9.1
black==25.9.0
click==8.3.0
Django==5.0.6
django-crispy-forms==2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
flake8==7.3.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.4.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytokens==0.1.10
pytz==2025.2
ruff==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
